Comments
- Single Line Comments                   - //
- Multi line Comments                     - /*   */
- XML Documentation Comments  -  ///

Comments are used to document what the program does and what specific
blocks or lines of code do. C# compiler ignores comments.

To Comment and Uncomment, there are 2 ways
1. Use the designer
2. Keyboard Shortcut: Ctrl+K, Ctrl+C, and Ctrl+K, Ctrl+U

Note: Don't try to comment every line of code. Use comments only for
blocks or lines of code that are difficult to understand