Datatypes Conversion
Implicit & Explicit conversion

Implicit Conversion is done by the compiler
	1. When there is no loss of information if the conversion is done
	2. If there is no possibility of throwing exception during the conversion

Example: Converting an int to a float will not lose any data and no exception will be thrown hence an implicit conversion can be done.

Whereas when converting a float to an int, we lose the fractional part and also a possibility of overflow exception. Hence, in this case an explicit conversion is required. For explicit conversion we use cast operator the convert class in C#