For & Foreach Loops
The for loop
A for loop is very similar to while loop. In a while loop we do the initialization at one place, condition check at another place, and modifying the variable at another place, whereas for loop has all of these at one place.

The foreach loop
A foreach loop is used to iterate through the items in a collection. For example, foreach loop can be used with arrays or collections such as ArrayList, HashTable and Generics