Introduction To Classes
What is a class?
We have seen data types like int, float, double etc. If you want to create complex custom types, then we can make use of classes.
A class consists of data and behaviour. Class data is represented by its fields and behaviour is represented by its methods.
Purpose of a class Constructor.
The purpose of a class constructor is to initialize class fields. A class constructor is automatically called when an instance of a class is created.
Constructors do not have return values and always have the same name as the class.
Constructors are not mandatory. If we do not provide a constructor, a default parameter less constructor is automatically provided.
Constructors can be overloaded by the number and type of parameters.
Destructors
Destructors have the same name as the class with ~ symbol in front of them
They don't take any parameters and do not return a value.
Destructors are places where you could put code to release any resources your class was holding during it