Difference between Classes and Structs
Structs are very similar to classes. We use struct keyword to create a struct.				
Just like classes structs can have fields, properties, constructors and methods.
However, there are several differences between classes and structs, which we will explore in this session.

Classes VS Structs
A struct is a value type where as a class is a reference type.
All the differences that are applicable to value types and reference types are also applicable to classes and structs.
Structs are stored on stack, whereas classes are stored on the heap.
Value types hold their value in memory where they are declared, but reference types hold a reference to an object in memory.
Value types are destroyed immediately after the scope is lost, whereas for reference types only the reference variable is destroyed after the scope is lost. The object is later destroyed by garbage collector.
When you copy a struct into another struct, a new copy of that struct gets created and modificatio